Move ViewPager-specific code into indicator library
This hides the implementation details from integrators (e.g. the mobile library).
